Change ipython console plugin imports after moving.
